backend/requirements.txt
flask
flask-cors
flask-socketio
python-dotenv
SQLAlchemy
PyMySQL
pygame
